+++ app/Http/Controllers/RucherController.php
@@ -2,10 +2,14 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\RucherRequest;
 use App\Models\Rucher;
-use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use App\Models\Exploitation;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use App\Http\Requests\RucherRequest;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Support\Facades\Gate;
 
 class RucherController extends Controller
 {
@@ -13,24 +17,6 @@ class RucherController extends Controller
 
     public function index(Request $request)
     {
-        // $user = $request->user();
-        // $role = $user->currentRole();
-
-        // if ($role->acces_complet_ruchers) {
-        //     $ruchers = $user->currentExploitation->ruchers()->withCount("ruches")->get();
-        // } else {
-        //     $ruchers = $role->ruchers()->wherePivot('peut_lire', '=', true)->withCount("ruches")->get();
-        // }
-        // return response()->json($ruchers);
-        // // $ruchers = $request->user()->currentExploitation->ruchers()->withCount("ruches")->get();
-
-        // if ($user->ownsCurrentExploitation()) {
-        //     $ruchers = $user->currentExploitation->ruchers()->withCount("ruches")->get();
-        // } else {
-        //     $ruchers = $user->currentRole()->ruchers()->wherePivot("peut_lire", "=", true)->withCount("ruches")->get();
-        //     $roles = $user->roles()->with('ruchers')->get();
-        //     $ruchers = $roles->pluck('ruchers')->flatten();
-        // }
         $role = $request->user()->currentRole();
         if (!$role) {
             return response()->json([]);
@@ -43,31 +29,56 @@ class RucherController extends Controller
     {
         $this->authorize('create', Rucher::class);
 
-        return Rucher::create($request->validated());
+        $rucher = DB::transaction(function () use ($request) {
+            $exploitationId = Auth::user()->current_exploitation_id;
+            $exploitation = Exploitation::findOrFail($exploitationId);
+            $rucher = $exploitation->ruchers()->create($request->validated());
+            $roles = $exploitation->roles()->get();
+            foreach ($roles as $role) {
+                if ($role->acces_complet_ruchers) {
+                    $rucher->roles()->attach($role->id, ["peut_modifier" => true, "peut_lire" => true]);
+                }
+            }
+            return $rucher;
+        });
+        return response()->json($rucher);
     }
 
     public function show(Rucher $rucher)
     {
         $this->authorize('view', $rucher);
-
         return $rucher;
     }
 
     public function update(RucherRequest $request, Rucher $rucher)
     {
         $this->authorize('update', $rucher);
-
         $rucher->update($request->validated());
-
         return $rucher;
     }
 
-    public function destroy(Rucher $rucher)
+    public function destroy(Request $request, int $rucherId)
     {
+        $validated = $request->validate([
+            "garder_ruches" => ["required", "boolean"],
+        ]);
+        $rucher = Rucher::find($rucherId);
+        if (!$rucher) {
+            abort(404, "Rucher introuvable.");
+        }
+        if (!$validated["garder_ruches"]) {
+            $rucher->ruches()->delete();
+            info("Suppression de toutes les ruches");
+        }
         $this->authorize('delete', $rucher);
-
         $rucher->delete();
+        return response()->json(["success" => true]);
+    }
 
-        return response()->json();
+    public function getRuches(Rucher $rucher)
+    {
+        Gate::authorize('view', $rucher);
+        $ruches = $rucher->ruches()->get();
+        return response()->json($ruches);
     }
 }
